video-intelligence-platform/video/backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
pydantic
pydantic-settings
openai
pinecone-client
langchain
langchain-openai
langchain-pinecone
langchain-core
langchain-community
langgraph
psycopg[binary]
psycopg-pool
moviepy
openai-whisper
flashrank
torch
numpy
requests
aiofiles